+++ src/main/java/DesignPatterns/StandardProblems/StrategyPatternDemo/PaymentStrategy.java
@@ -1,5 +1,5 @@
 package DesignPatterns.StandardProblems.StrategyPatternDemo;
 
 public interface PaymentStrategy {
-    public void pay(int amount);
+    public void pay(double amount);
 }
